rename test component
