+++ tests/components/ct-vue-vite/src/components/SlotDefaultValue.vue
